extract components from Alert
